+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Add change column to Diff transformation providing column names of all changed columns in a row.
+- Add Python API for Diff transformation.
 
 ## [1.0.0] - 2020-03-12